feature(setup): added `walletkit-ui`
