+++ .github/workflows/ci.yml
@@ -46,10 +46,9 @@ jobs:
       - run: pnpm install --frozen-lockfile
       - run: pnpm test -- -- --maxWorkers=100% --coverage
 
-  # TODO(thedoublejay): add when there are test cases
-  #      - uses: codecov/codecov-action@d9f34f8cd5cb3b3eb79b3e4b5dae3a16df499a70 # tag=v3.1.1
-  #        with:
-  #          fail_ci_if_error: true
+      - uses: codecov/codecov-action@d9f34f8cd5cb3b3eb79b3e4b5dae3a16df499a70 # tag=v3.1.1
+        with:
+          fail_ci_if_error: true
 
   lint_prettier:
     name: Lint (prettier)
