+++ package.json
@@ -16,6 +16,7 @@
   "devDependencies": {
     "@birthdayresearch/eslint-config": "^0.3.2",
     "@birthdayresearch/sticky-turbo": "^0.3.2",
+    "@waveshq/walletkit-core": "workspace:0.0.0",
     "@waveshq/standard-web": "^0.15.1",
     "@waveshq/standard-web-linter": "^0.15.1",
     "@waveshq/standard-defichain-jellyfishsdk": "^0.15.1",
@@ -27,6 +28,9 @@
     ],
     "parserOptions": {
       "project": "./tsconfig.json"
+    },
+    "rules": {
+      "check-file/filename-naming-convention": "off"
     }
   },
   "lint-staged": {
