+++ packages/walletkit-core/package.json
@@ -12,7 +12,8 @@
   },
   "scripts": {
     "build": "tsc -b ./tsconfig.build.json",
-    "test": "jest --maxWorkers=100%"
+    "test": "jest --maxWorkers=100%",
+    "lint": "eslint ."
   },
   "jest": {
     "preset": "@birthdayresearch/sticky-jest"
