+++ packages/walletkit-core/src/api/environment.unit.ts
@@ -24,29 +24,29 @@ describe("environments", () => {
 
     expect(env.name).toStrictEqual(EnvironmentName.Development);
   });
-});
 
-it("should match development for default", () => {
-  const env = getEnvironment("development");
-  expect(env.name).toStrictEqual(EnvironmentName.Development);
-});
+  it("should match development for default", () => {
+    const env = getEnvironment("development");
+    expect(env.name).toStrictEqual(EnvironmentName.Development);
+  });
 
-it("should match Remote Playground on Production environment", () => {
-  const env = getEnvironment("production");
-  expect(env.name).toStrictEqual(EnvironmentName.Production);
-  // To be changed to MainNet on launch
-  expect(env.networks[0]).toStrictEqual(EnvironmentNetwork.MainNet);
-});
+  it("should match Remote Playground on Production environment", () => {
+    const env = getEnvironment("production");
+    expect(env.name).toStrictEqual(EnvironmentName.Production);
+    // To be changed to MainNet on launch
+    expect(env.networks[0]).toStrictEqual(EnvironmentNetwork.MainNet);
+  });
 
-it("should match Remote Playground on prerelease environment", () => {
-  const env = getEnvironment("prerelease");
-  expect(env.name).toStrictEqual(EnvironmentName.Preview);
-  expect(env.networks[0]).toStrictEqual(EnvironmentNetwork.RemotePlayground);
-});
+  it("should match Remote Playground on prerelease environment", () => {
+    const env = getEnvironment("prerelease");
+    expect(env.name).toStrictEqual(EnvironmentName.Preview);
+    expect(env.networks[0]).toStrictEqual(EnvironmentNetwork.RemotePlayground);
+  });
 
-it("should check if isPlayground", () => {
-  expect(isPlayground(EnvironmentNetwork.LocalPlayground)).toBe(true);
-  expect(isPlayground(EnvironmentNetwork.RemotePlayground)).toBe(true);
-  expect(isPlayground(EnvironmentNetwork.MainNet)).toBe(false);
-  expect(isPlayground(EnvironmentNetwork.TestNet)).toBe(false);
+  it("should check if isPlayground", () => {
+    expect(isPlayground(EnvironmentNetwork.LocalPlayground)).toBe(true);
+    expect(isPlayground(EnvironmentNetwork.RemotePlayground)).toBe(true);
+    expect(isPlayground(EnvironmentNetwork.MainNet)).toBe(false);
+    expect(isPlayground(EnvironmentNetwork.TestNet)).toBe(false);
+  });
 });
