+++ packages/walletkit-ui/README.md
@@ -1,0 +1,5 @@
+[![npm](https://img.shields.io/npm/v/@waveshq/walletkit-ui)](https://www.npmjs.com/package/@waveshq/walletkit-ui/v/latest)
+
+# @waveshq/walletkit-ui
+
+React components for building wallet
