+++ packages/walletkit-ui/src/contexts/logger.ts
@@ -1,0 +1,4 @@
+export interface BaseLogger {
+  error: (error: any) => void;
+  info: (message: string) => void;
+}
